+++ OUDS/Core/Tokens/SemanticTokens/Sources/Values/ColorMultipleSemanticTokens.swift
@@ -296,7 +296,7 @@ public protocol ColorMultipleSemanticTokens {
 
     var colorChartCategoricalTier10: MultipleColorSemanticTokens { get }
 
-    var colorChartFunctionalInformation: MultipleColorSemanticTokens { get }
+    var colorChartFunctionalInfo: MultipleColorSemanticTokens { get }
 
     var colorChartFunctionalNegative: MultipleColorSemanticTokens { get }
 
@@ -358,6 +358,8 @@ public protocol ColorMultipleSemanticTokens {
 
     var colorRepositoryNeutralMutedWhite: MultipleColorSemanticTokens { get }
 
+    var colorRepositoryOpacityBlackHigh: MultipleColorSemanticTokens { get }
+
     var colorRepositoryOpacityBlackHigher: MultipleColorSemanticTokens { get }
 
     var colorRepositoryOpacityBlackHighest: MultipleColorSemanticTokens { get }
@@ -370,11 +372,9 @@ public protocol ColorMultipleSemanticTokens {
 
     var colorRepositoryOpacityBlackMedium: MultipleColorSemanticTokens { get }
 
+    var colorRepositoryOpacityBlackSoft: MultipleColorSemanticTokens { get }
+
     var colorRepositoryOpacityBlackTransparent: MultipleColorSemanticTokens { get }
-
-    var colorRepositoryOpacityWhiteHigher: MultipleColorSemanticTokens { get }
-
-    var colorRepositoryOpacityWhiteHigh: MultipleColorSemanticTokens { get }
 
     var colorRepositoryOpacityInfo: MultipleColorSemanticTokens { get }
 
@@ -384,7 +384,13 @@ public protocol ColorMultipleSemanticTokens {
 
     var colorRepositoryOpacityWarning: MultipleColorSemanticTokens { get }
 
+    var colorRepositoryOpacityWhiteHigh: MultipleColorSemanticTokens { get }
+
+    var colorRepositoryOpacityWhiteHigher: MultipleColorSemanticTokens { get }
+
     var colorRepositoryOpacityWhiteHighest: MultipleColorSemanticTokens { get }
+
+    var colorRepositoryOpacityWhiteMedium: MultipleColorSemanticTokens { get }
 
     var colorRepositoryOpacityWhiteLow: MultipleColorSemanticTokens { get }
 
@@ -395,6 +401,8 @@ public protocol ColorMultipleSemanticTokens {
     var colorRepositoryOpacityWhiteTransparent: MultipleColorSemanticTokens { get }
 
     var colorRepositoryPositiveDefault: MultipleColorSemanticTokens { get }
+
+    var colorRepositoryPositiveHigh: MultipleColorSemanticTokens { get }
 
     var colorRepositoryPositiveHighest: MultipleColorSemanticTokens { get }